#16 moved login logic to application controller, now login only shows a message, also redirecting away from login route if user is already authenticated
