+++ app/routes/login.js
@@ -1,4 +1,10 @@
 import Ember from 'ember';
 
 export default Ember.Route.extend({
+  session: Ember.inject.service(),
+  beforeModel() {
+    if(this.get('session.isAuthenticated')) {
+      this.transitionTo('dashboard');
+    }
+  }
 });
