+++ app/templates/login.hbs
@@ -1,1 +1,1 @@
-<a class="btn btn-success" onclick={{action 'login'}}>Login</a>
+<h2 class="text-center">You need to login to access the app</h2>
